+++ src/bin/restart.sh
@@ -1,0 +1,6 @@
+#! /bin/shell
+
+echo stop fast-spring-boot Application...
+sh stop.sh
+echo start fast-spring-boot Application...
+sh start.sh