+++ src/main/java/io/geekidea/fastspringboot/system/mapper/SysLogMapper.java
@@ -17,7 +17,7 @@ import java.io.Serializable;
  * </p>
  *
  * @author liujixiang
- * @since 2018-11-10
+ * @since 2018-11-30
  */
 @Repository
 public interface SysLogMapper extends BaseMapper<SysLog> {