+++ src/main/java/io/geekidea/fastspringboot/system/web/controller/SysLogController.java
@@ -1,6 +1,5 @@
 package io.geekidea.fastspringboot.system.web.controller;
 
-import com.alibaba.fastjson.JSON;
 import io.geekidea.fastspringboot.system.entity.SysLog;
 import io.geekidea.fastspringboot.system.service.SysLogService;
 import io.geekidea.fastspringboot.system.web.param.SysLogQueryParam;
@@ -18,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.validation.Valid;
 
 /**
  * <p>
@@ -25,7 +25,7 @@ import org.springframework.web.bind.annotation.RestController;
  * </p>
  *
  * @author liujixiang
- * @since 2018-11-10
+ * @since 2018-11-30
  */
 @RestController
 @RequestMapping("/sysLog")
@@ -41,10 +41,8 @@ public class SysLogController extends BaseController {
     */
     @PostMapping("/add")
     @ApiOperation(value = "添加SysLog对象",notes = "添加系统日志",response = ApiResult.class)
-    public ApiResult<Boolean> addSysUser(@RequestBody SysLog sysLog) throws Exception{
-        log.debug("sysLog:{}", JSON.toJSONString(sysLog));
+    public ApiResult<Boolean> addSysUser(@Valid @RequestBody SysLog sysLog) throws Exception{
         boolean flag = sysLogService.save(sysLog);
-        log.debug("flag:{}",flag);
         return success(flag);
     }
 
@@ -53,10 +51,8 @@ public class SysLogController extends BaseController {
     */
     @PostMapping("/update")
     @ApiOperation(value = "修改SysLog对象",notes = "修改系统日志",response = ApiResult.class)
-    public ApiResult<Boolean> updateSysUser(@RequestBody SysLog sysLog) throws Exception{
-        log.debug("sysLog:{}", JSON.toJSONString(sysLog));
+    public ApiResult<Boolean> updateSysUser(@Valid @RequestBody SysLog sysLog) throws Exception{
         boolean flag = sysLogService.updateById(sysLog);
-        log.debug("flag:{}",flag);
         return success(flag);
     }
 
@@ -65,10 +61,8 @@ public class SysLogController extends BaseController {
     */
     @PostMapping("/delete")
     @ApiOperation(value = "删除SysLog对象",notes = "删除系统日志",response = ApiResult.class)
-    public ApiResult<Boolean> deleteSysUser(@RequestBody IdParam idParam) throws Exception{
-        log.debug("idParam:{}", JSON.toJSONString(idParam));
+    public ApiResult<Boolean> deleteSysUser(@Valid @RequestBody IdParam idParam) throws Exception{
         boolean flag = sysLogService.removeById(idParam.getId());
-        log.debug("flag:{}",flag);
         return success(flag);
     }
 
@@ -77,10 +71,8 @@ public class SysLogController extends BaseController {
     */
     @PostMapping("/info")
     @ApiOperation(value = "获取SysLog对象详情",notes = "查看系统日志",response = SysLogQueryVo.class)
-    public ApiResult<SysLogQueryVo> getSysUser(@RequestBody IdParam idParam) throws Exception{
-        log.debug("idParam:{}", JSON.toJSONString(idParam));
+    public ApiResult<SysLogQueryVo> getSysUser(@Valid @RequestBody IdParam idParam) throws Exception{
         SysLogQueryVo sysLogQueryVo = sysLogService.getSysLogById(idParam.getId());
-        log.debug("sysLogQueryVo:{}",sysLogQueryVo);
         return success(sysLogQueryVo);
     }
 
@@ -89,10 +81,8 @@ public class SysLogController extends BaseController {
      */
     @PostMapping("/getPageList")
     @ApiOperation(value = "获取SysLog分页列表",notes = "系统日志分页列表",response = SysLogQueryVo.class)
-    public ApiResult<Paging<SysLogQueryVo>> getSysLogPageList(@RequestBody(required = false) SysLogQueryParam sysLogQueryParam) throws Exception{
-        log.debug("sysLogQueryParam:{}", JSON.toJSONString(sysLogQueryParam));
+    public ApiResult<Paging<SysLogQueryVo>> getSysLogPageList(@Valid @RequestBody(required = false) SysLogQueryParam sysLogQueryParam) throws Exception{
         Paging<SysLogQueryVo> paging = sysLogService.getSysLogPageList(sysLogQueryParam);
-        log.debug("paging:{}",paging);
         return success(paging);
     }