+++ src/main/java/io/geekidea/fastspringboot/system/web/param/SysLogQueryParam.java
@@ -11,7 +11,7 @@ import lombok.EqualsAndHashCode;
  * </p>
  *
  * @author liujixiang
- * @date 2018-11-10
+ * @date 2018-11-30
  */
 @Data
 @EqualsAndHashCode(callSuper = true)